Add support for kth_space_lab world and gazebo-ros bridge in launch files
